scrapers/pyproject.toml
[project]
name = "nfl-de-scrapers"
version = "0.2.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.3.0",
    "supabase>=2.8.0",
    "python-dotenv>=1.0.1",
    "tenacity>=9.0.0",
    "feedparser>=6.0.11",
    "deep-translator>=1.11.4",
]
